+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## [Unreleased]
 ### Added
-- ui file for MainWindow
+- Basic model.